Refactor: Update `getPaddingBottom` ratio parameter type
The `ratio` parameter in the `UIUtils.getPaddingBottom` method has been changed from `int` to `double`.

This allows for more granular control over the padding calculation, enabling fractional ratios if needed.

+++ lib/mckimquyen/util/ui_utils.dart
@@ -1028,7 +1028,7 @@ class UIUtils {
     }
   }
 
-  static double getPaddingBottom(BuildContext context, {int ratio = 2}) {
+  static double getPaddingBottom(BuildContext context, {double ratio = 2}) {
     final padding = MediaQuery.of(context).padding;
     return padding.bottom * ratio;
   }
